Remove policy_id and statuses param and implement querying db for a particular rule_id
