+++ packages/mds-compliance/api.ts
@@ -20,8 +20,7 @@ import stream from 'mds-stream'
 import db from 'mds-db'
 import jwtDecode from 'jwt-decode'
 import log from 'mds-logger'
-import { EVENT_STATUS_MAP, VEHICLE_STATUS } from 'mds-enums'
-import { isUUID, now, days, pathsFor, head, getPolygon, pointInShape } from 'mds-utils'
+import { isUUID, now, days, pathsFor, head, getPolygon, pointInShape, isInStatesOrEvents } from 'mds-utils'
 import { Policy, Geography, VehicleEvent, ComplianceResponse, Device, UUID } from 'mds'
 import { Geometry, FeatureCollection } from 'geojson'
 import * as compliance_engine from './mds-compliance-engine'
@@ -239,7 +238,7 @@ function api(app: express.Express): express.Express {
     }
   })
 
-  app.get(pathsFor('/count/:policy_id/:rule_id'), async (req: express.Request, res: express.Response) => {
+  app.get(pathsFor('/count/:rule_id'), async (req: express.Request, res: express.Response) => {
     if (
       !['5f7114d1-4091-46ee-b492-e55875f7de00', '45f37d69-73ca-4ca6-a461-e7283cffa01a'].includes(res.locals.provider_id)
     ) {
@@ -253,14 +252,9 @@ function api(app: express.Express): express.Express {
       })
     }
 
-    const { policy_id, rule_id } = req.params
-    let { statuses } = req.query
-    statuses = statuses ? (JSON.parse(statuses) as VEHICLE_STATUS[]) : undefined
+    const { rule_id } = req.params
     try {
-      const [policy] = await db.readPolicies({ policy_id })
-      const rule = policy.rules.find(r => {
-        return r.rule_id === rule_id
-      })
+      const rule = await db.readRule(rule_id)
       if (!rule) {
         throw new Error('Rule not found in specified Policy')
       }
@@ -280,10 +274,7 @@ function api(app: express.Express): express.Express {
         return [...acc, getPolygon(geographies, geography.geography_id)]
       }, [])
 
-      // const devices = (await db.readDeviceIds()).map((record: { device_id: UUID; provider_id: UUID }) => record.device_id)
-      const events = statuses
-        ? (await cache.readAllEvents()).filter(event => statuses.includes(EVENT_STATUS_MAP[event.event_type]))
-        : await cache.readAllEvents()
+      const events = (await cache.readAllEvents()).filter(event => isInStatesOrEvents(rule, event))
 
       const count = events.reduce((count_acc, event) => {
         return (
